// src/app/models/IData.ts
export interface IData {
  id: string;
  information1?: string;
  information2?: string;
  information3?: string;
  information4?: string;
  information5?: string;
  information6?: string;
  information7?: string;
  information8?: string;
  information9?: string;
  information10?: string;
  information11?: string;
  information12?: string;
  information13?: string;
  information14?: string;
  information15?: string;
}